requirements.txt
asgiref==3.7.2
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.0
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.0.0
platformdirs==3.5.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.0
